# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Full Stack",
        "version": "0.1"
    },
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "operationId": "6edfb1b913bd68fb21ed7bdbe58ddc3f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "c_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Cristiano Mendes",
                                    "email": "cristiano@mendes.dev",
                                    "password": "Password@123",
                                    "c_password": "Password@123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true,
                                            "data": {
                                                "token": "1|5H4gXtymF8LlJkAghbSPLMoo8vNXL5sbGO3FN0iC3e03fb26",
                                                "user": {
                                                    "id": "user-bf2-fcb7-4ab5-8e38-e60fb3766306",
                                                    "name": "Cristiano Mendes",
                                                    "email": "cristiano@zenus.app"
                                                }
                                            },
                                            "message": "Registrado com sucesso."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in",
                "description": "Sign in",
                "operationId": "e6e5a6e33cd0fa60f7d959ab9e4a1900",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "cristiano@mendes.dev",
                                    "password": "Password@123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true,
                                            "data": {
                                                "token": "1|5H4gXtymF8LlJkAghbSPLMoo8vNXL5sbGO3FN0iC3e03fb26",
                                                "user": {
                                                    "id": "user-bf2-fcb7-4ab5-8e38-e60fb3766306",
                                                    "name": "Cristiano Mendes",
                                                    "email": "cristiano@zenus.app"
                                                }
                                            },
                                            "message": "Registrado com sucesso."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign out",
                "description": "Sign out",
                "operationId": "c7042c8f65b3b22cc6e1a159b8381b9a",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true,
                                            "data": "[]",
                                            "message": "Você foi deslogado com sucesso."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/balance": {
            "get": {
                "tags": [
                    "Balance"
                ],
                "summary": "Get Logged User Balance",
                "operationId": "cba0391080557167de89d3cda0c823fa",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true,
                                            "data": {
                                                "id": "balance--38db-44e9-84ef-31a38419d788",
                                                "user_id": "user-bf2-fcb7-4ab5-8e38-e60fb3766306",
                                                "value": "0.00"
                                            },
                                            "message": "Saldo obtido com sucesso."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Logged User Statement",
                "operationId": "8a8ad40cfb14c57f0740a61d274e6b3f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true,
                                            "data": "[]",
                                            "message": "Extrato obtido com sucesso."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/deposit": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Make a deposit",
                "description": "Make a deposit",
                "operationId": "8f8dae757df3daea80711bda92e774cf",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "value": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "value": 10
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true,
                                            "data": {
                                                "id": "transaction-a-437b-b40d-0c0cb0c340e5",
                                                "user_id": "user-f91-6bd1-45db-9eb7-75dbf7bec959",
                                                "user": "Desenvolvimento",
                                                "payer_id": "user-f91-6bd1-45db-9eb7-75dbf7bec959",
                                                "payer": "Desenvolvimento",
                                                "type": "Entrada",
                                                "category": "Depósito",
                                                "value": "10,00",
                                                "description": "Depósito em conta.",
                                                "created_at": "11/11/2024 16:55:23"
                                            },
                                            "message": "Depósito efetuado com sucesso."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transfer": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Send value to another user",
                "description": "Send value to another user",
                "operationId": "e12606a4edf37fc8d7cb4bc0c15ad89e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "uuid"
                                    },
                                    "value": {
                                        "type": "number"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "value": 10,
                                    "user_id": "user-f91-6bd1-45db-9eb7-75dbf7bec959",
                                    "date": "2024-11-11"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true,
                                            "data": {
                                                "id": "transaction-1-47d8-8ede-4b25c916cc92",
                                                "user_id": "user-f91-6bd1-45db-9eb7-75dbf7bec959",
                                                "user": "Desenvolvimento",
                                                "payer_id": "user-f91-6bd1-45db-9eb7-75dbf7bec959",
                                                "payer": "Desenvolvimento",
                                                "type": "Saía",
                                                "category": "Transferência",
                                                "value": "5,00",
                                                "description": "Transferência para conta de Desenvolvimento",
                                                "created_at": "11/11/2024 17:40:38"
                                            },
                                            "message": "Transferência efetuada com sucesso."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/{id}": {
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Cancel a transaction",
                "operationId": "54e54f926db0f3353d829ba0eaba819c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the transaction",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "An UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Balance",
            "description": "Balance"
        },
        {
            "name": "Transactions",
            "description": "Transactions"
        }
    ]
}